requirements.txt
matplotlib
scipy
tensorflow-gpu
numpy
Pillow
